Project finished, edits done to index.html, style.css and script.js
index.html: Added H1 in header, added banner to anounce winner, added
container for the image

style.css: added id for h1 in header and for the announcement banner.
added new font. Added CSS for the image container and image.

script.js: Score is now recorded on the webpage, image pops up on the
webpage when a round is played showing what each player played.
Functions added:
computerDisplay and userDisplay  clears and adds image
of played hand to the html.
checkWinner: Checks who won the game and outputs the winner to the
banner.
resetGame: clears the page of score, round, image and banner.

moved constants that were in function scope to Global scope as they
were used in multiple functions.

added eventlistner for when user clicks cancel button.